added rate for book

+++ lib/src/book/presentation/blocs/book_detail/book_detail_event.dart
@@ -17,7 +17,7 @@ class OnSetBookDetailEvent extends BookDetailEvent {
 }
 
 class OnRateBookEvent extends BookDetailEvent {
-  final int rate;
+  final double rate;
 
   const OnRateBookEvent(this.rate);
